Move DisplayData out of graph package
